+++ backend/src/main/java/pwr/piisw/backend/dto/ChanThreadDto.java
@@ -1,0 +1,17 @@
+package pwr.piisw.backend.dto;
+
+import java.time.LocalDateTime;
+
+import pwr.piisw.backend.models.ChanThreadCategory;
+
+import lombok.*;
+
+@Data
+public class ChanThreadDto {
+  private Integer id;
+  private ChanThreadCategory category;
+  private String nickname;
+  private String content;
+  private String imgUrl;
+  private LocalDateTime date;
+}