+++ backend/src/main/java/pwr/piisw/backend/dto/ImageDto.java
@@ -1,0 +1,9 @@
+package pwr.piisw.backend.dto;
+
+import lombok.Data;
+
+@Data
+public class ImageDto {
+  private String id;
+  private String imgUrl;
+}