+++ backend/src/main/java/pwr/piisw/backend/exceptions/ThreadNotFoundException.java
@@ -1,0 +1,7 @@
+package pwr.piisw.backend.exceptions;
+
+public class ThreadNotFoundException extends RuntimeException {
+  public ThreadNotFoundException(String message) {
+    super(message);
+  }
+}