+++ backend/src/main/java/pwr/piisw/backend/helper/OffsetBasedPageRequest.java
@@ -1,0 +1,76 @@
+package pwr.piisw.backend.helper;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+
+public class OffsetBasedPageRequest implements Pageable {
+
+  private int limit;
+  private int offset;
+  private String sortBy;
+  private Sort.Direction direction;
+
+  public OffsetBasedPageRequest(int limit, int offset, String sortBy, Sort.Direction direction) {
+    if (limit < 1) {
+      throw new IllegalArgumentException("Limit must not be less than one!");
+    }
+    if (offset < 0) {
+      throw new IllegalArgumentException("Offset index must not be less than zero!");
+    }
+    this.limit = limit;
+    this.offset = offset;
+    this.sortBy = sortBy;
+    this.direction = direction;
+  }
+
+  @Override
+  public int getPageNumber() {
+    return offset / limit;
+  }
+
+  @Override
+  public int getPageSize() {
+    return limit;
+  }
+
+  @Override
+  public long getOffset() {
+    return offset;
+  }
+
+  @Override
+  public Sort getSort() {
+    return Sort.by(direction, sortBy);
+  }
+
+  @Override
+  public Pageable next() {
+    // Typecast possible because number of entries cannot be bigger than integer (primary key is
+    // integer)
+    return new OffsetBasedPageRequest(
+        getPageSize(), (int) (getOffset() + getPageSize()), sortBy, direction);
+  }
+
+  public Pageable previous() {
+    // The integers are positive. Subtracting does not let them become bigger than integer.
+    return hasPrevious()
+        ? new OffsetBasedPageRequest(
+            getPageSize(), (int) (getOffset() - getPageSize()), sortBy, direction)
+        : this;
+  }
+
+  @Override
+  public Pageable previousOrFirst() {
+    return hasPrevious() ? previous() : first();
+  }
+
+  @Override
+  public Pageable first() {
+    return new OffsetBasedPageRequest(getPageSize(), 0, sortBy, direction);
+  }
+
+  @Override
+  public boolean hasPrevious() {
+    return offset > limit;
+  }
+}