+++ backend/src/main/java/pwr/piisw/backend/models/ChanThreadCategory.java
@@ -1,0 +1,14 @@
+package pwr.piisw.backend.models;
+
+public enum ChanThreadCategory {
+  IT,
+  MATH,
+  MED,
+  BIO,
+  CHEM,
+  PHYS,
+  ELEC,
+  BUD,
+  ARCH,
+  MECH;
+}