+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [[0.9.0]] - 2021-06-20
+
+### Added
+
+- Backend model
+- Backend REST Controller
+- Backend limit and offset
+- Backend endpoints
+- Database
+
 ## [[0.8.3]] - 2021-06-15
 
 ### Added
@@ -177,6 +187,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `README.md`
 
+[0.9.0]: https://github.com/Luzkan/Polichan/releases/tag/v0.9.0
 [0.8.3]: https://github.com/Luzkan/Polichan/releases/tag/v0.8.3
 [0.8.2]: https://github.com/Luzkan/Polichan/releases/tag/v0.8.2
 [0.8.1]: https://github.com/Luzkan/Polichan/releases/tag/v0.8.1